// backend-cluster/backend-v2/src/features/gitea/client/gitea-api.ts
export interface ChangedFile {
  /** @format int64 */
  additions?: number;
  /** @format int64 */
  changes?: number;
  contents_url?: string;
  /** @format int64 */
  deletions?: number;
  filename?: string;
  html_url?: string;
  previous_filename?: string;
  raw_url?: string;
  status?: string;
}
// ...
export interface ContentsResponse {
  /** FileLinksResponse contains the links for a repo's file */
  _links?: FileLinksResponse;
  /** `content` is populated when `type` is `file`, otherwise null */
  content?: string;
  download_url?: string;
  /** `encoding` is populated when `type` is `file`, otherwise null */
  encoding?: string;
  git_url?: string;
  html_url?: string;
  /** @format date-time */
  last_author_date?: string;
  last_commit_sha?: string;
  /** @format date-time */
  last_committer_date?: string;
  name?: string;
  path?: string;
  sha?: string;
  /** @format int64 */
  size?: number;
  /** `submodule_git_url` is populated when `type` is `submodule`, otherwise null */
  submodule_git_url?: string;
  /** `target` is populated when `type` is `symlink`, otherwise null */
  target?: string;
  /** `type` will be `file`, `dir`, `symlink`, or `submodule` */
  type?: string;
  url?: string;
}
// ...
export interface FileLinksResponse {
  git?: string;
  html?: string;
  self?: string;
}